Move construction to utils
