+++ py-polars/polars/series/series.py
@@ -19,14 +19,6 @@ from typing import (
 )
 
 from polars import internals as pli
-from polars._construction import (
-    arrow_to_pyseries,
-    iterable_to_pyseries,
-    numpy_to_pyseries,
-    pandas_to_pyseries,
-    sequence_to_pyseries,
-    series_to_pyseries,
-)
 from polars.datatypes import (
     Boolean,
     Categorical,
@@ -71,6 +63,14 @@ from polars.series.string import StringNameSpace
 from polars.series.struct import StructNameSpace
 from polars.series.utils import expr_dispatch, get_ffi_func
 from polars.slice import PolarsSlice
+from polars.utils._construction import (
+    arrow_to_pyseries,
+    iterable_to_pyseries,
+    numpy_to_pyseries,
+    pandas_to_pyseries,
+    sequence_to_pyseries,
+    series_to_pyseries,
+)
 from polars.utils.convert import (
     _date_to_pl_date,
     _datetime_to_pl_timestamp,
