+++ py-polars/tests/unit/test_df.py
@@ -14,7 +14,6 @@ import pyarrow as pa
 import pytest
 
 import polars as pl
-from polars._construction import iterable_to_pydf
 from polars.datatypes import DTYPE_TEMPORAL_UNITS, INTEGER_DTYPES
 from polars.testing import (
     assert_frame_equal,
@@ -22,6 +21,7 @@ from polars.testing import (
     assert_series_equal,
 )
 from polars.testing.parametric import columns
+from polars.utils._construction import iterable_to_pydf
 
 if TYPE_CHECKING:
     from polars.type_aliases import JoinStrategy, UniqueKeepStrategy
