+++ py-polars/tests/unit/test_series.py
@@ -10,7 +10,6 @@ import pyarrow as pa
 import pytest
 
 import polars as pl
-from polars._construction import iterable_to_pyseries
 from polars.datatypes import (
     Date,
     Datetime,
@@ -24,6 +23,7 @@ from polars.datatypes import (
 )
 from polars.exceptions import ShapeError
 from polars.testing import assert_frame_equal, assert_series_equal
+from polars.utils._construction import iterable_to_pyseries
 
 if TYPE_CHECKING:
     from polars.type_aliases import EpochTimeUnit, TimeUnit
